export/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moveread-export"
version = "0.1.0"
authors = [
  {name="Marcel Claramunt", email="marcel@moveread.com"}
]
description = "Exporting data from annotations of the Moveread Core"
dependencies = [
  "haskellian-either", "pure-cv",
  "moveread-boxes", "moveread-labels", "moveread-errors"
]
requires-python = ">=3.10"
readme = {file="README.md", content-type="text/markdown"}

# [project.urls]
# repo = "<GIT_REPO_URL>"

# [project.optional-dependencies]
# test = [
#   "pytest < 5.0.0",
#   "pytest-cov[all]"
# ]